+++ src/Extensions/RssCore/SkipDay.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Nexendrie\Rss\Extensions\RssCore;
+
+enum SkipDay {
+  case Monday;
+  case Tuesday;
+  case Wednesday;
+  case Thursday;
+  case Friday;
+  case Saturday;
+  case Sunday;
+}
+
+?>